Recipe for geminabox server
Based on a ruby 2.3 image, add the source for geminabox 0.13.1
and install all the dependencies and launch the rack server.

Note: using a submodule for geminabox

Need to test uploading gems to the server, and using gems with
the server as a source (and doing that from a different container).

Still need to make the data a long-lived shared volume .. that's
the next step to do after understanding how geminabox works.