Updated to persist geminabox store
Added a ruby venv to run 'gem inabox' commands.
Added a README to describe how to persist the gem source gems.
Added a README to describe how to add a data disk to a boot2docker VM.
Changed the docker run command to add a data volume for the
geminabox server.