Changing to use the ruby 2.3 'alpine' image
Alpine Linux uses 'apk' instead of 'apt-get' and 'adduser' instead
of 'useradd' and users get the 'ash' shell instead of the 'bash' shell.

And Alpine uses the 'musl' version of glibc, so there could be
some packages that aren't supported by 'apk' and 'musl'.

But it is a much smaller Docker image.